Cambios de la Auth, los diseños de las vistas y la organizacion de carpetas -mhp
