+++ lib/main.dart
@@ -1,9 +1,13 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:proyectomovil/app.dart';
+import 'package:get/get.dart';
+import 'package:get/get_core/src/get_main.dart';
+import 'package:proyectomovil/ui/screens/app.dart';
+import 'package:proyectomovil/domain/controllers/gestionUser/controlleruser.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
+  Get.put(ControlUserAuth());
   runApp(const MyApp());
 }
