+++ lib/ui/screens/gestionInventario/Inventario/consultarInventario.dart
