+++ lib/ui/screens/gestionInventario/Inventario/detalleMedicamento.dart
