+++ lib/ui/screens/screens.dart
@@ -1,14 +1,14 @@
-export 'package:proyectomovil/ui/screens/home/panel_De_Control.dart';
+export 'package:proyectomovil/ui/screens/panelDeControl/panel_De_Control.dart';
 export 'package:proyectomovil/ui/screens/home/home.dart';
-export 'package:proyectomovil/ui/screens/gestionUser/registrar.dart';
+export 'package:proyectomovil/ui/screens/gestionUser/register.dart';
 export 'package:proyectomovil/ui/screens/gestionUser/login.dart';
-export 'package:proyectomovil/ui/screens/gestionPacientes/ingresarDatosPropietario.dart';
-export 'package:proyectomovil/ui/screens/gestionPacientes/ingresarPacientes.dart';
-export 'package:proyectomovil/ui/screens/gestionInventario/IngresarCuadroClinicoParte1.dart';
-export 'package:proyectomovil/ui/screens/gestionInventario/IngresarCuadroClinicoParte2.dart';
-export 'package:proyectomovil/ui/screens/gestionInventario/consultarMedicamentos.dart';
-export 'package:proyectomovil/ui/screens/gestionInventario/infoMedicamentos.dart';
-export 'package:proyectomovil/ui/screens/gestionInventario/ingresarMedicamentos.dart';
+export 'package:proyectomovil/ui/screens/gestionPacientes/ingreasarPacientes/ingresarDatosPropietario.dart';
+export 'package:proyectomovil/ui/screens/gestionPacientes/ingreasarPacientes/ingresarPacientes.dart';
+export 'package:proyectomovil/ui/screens/gestionPacientes/ingreasarPacientes/IngresarCuadroClinicoParte2.dart';
+export 'package:proyectomovil/ui/screens/gestionPacientes/ingreasarPacientes/IngresarCuadroClinicoParte1.dart';
+export 'package:proyectomovil/ui/screens/gestionInventario/Inventario/consultarMedicamentos.dart';
+export 'package:proyectomovil/ui/screens/gestionInventario/Inventario/infoMedicamentos.dart';
+export 'package:proyectomovil/ui/screens/gestionInventario/ingresarMedicamento/ingresarMedicamentos.dart';
