+++ pubspec.yaml
@@ -22,17 +22,30 @@ environment:
   sdk: '>=3.1.0 <4.0.0'
 
 dependencies:
-  cloud_firestore: ^4.12.1
-  cupertino_icons: ^1.0.2
-  firebase_auth: ^4.12.0
-  firebase_core: ^2.17.0
-  firebase_storage: ^11.4.0
+
+  # cupertino_icons: ^1.0.2
+  firebase_auth: ^4.12.1
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^10.5.0
+
+  cupertino_icons: ^1.0.2
   get: ^4.6.6
+  firebase_core: ^2.20.0
   get_storage: ^2.1.1
-  passwordfield: ^0.2.0
+  image_picker: ^1.0.4
+  cloud_firestore: ^4.12.2
+  firebase_storage: ^11.4.1
+  font_awesome_flutter: ^10.5.0
+  flutter_svg: ^2.0.9
+ 
+  
+  # get: ^4.6.6
+  # get_storage: ^2.1.1
+  # passwordfield: ^0.2.0
+  # firebase_core: ^2.20.0
+  # image_picker: ^1.0.4
+  # cloud_firestore: ^4.12.2
+  # firebase_storage: ^11.4.1
 
 dev_dependencies:
   flutter_lints: ^2.0.0
